+++ CHANGES.md
@@ -2,6 +2,25 @@
 
 Changes for each version along with any requirements to update from the previous version will be listed below.
 
+## Minor Version Ensign – 1 – 0 Update
+
+### Fleet Stats
+
+* Added a member details modal with:
+    * A timeline of Position in Fleet, Ships Used, and Systems Visited
+    * An Event Log of the aforementioned
+
+### Bug Fixes
+
+* Fixed an asynchronous issue with tracking multiple shared fleets.
+* Fixed an issue where the tracking page would fail to track a fleet with the status "Starting". 
+
+### UPDATE INSTRUCTIONS (From Version Ensign – 0 – *)
+
+1. Pause operation of the Fleet Checker.
+2. Sync up files with the repository.
+3. Restart operation of the Fleet Checker.
+
 ## Patch Version Ensign – 0 – 1 Update
 
 ### Tracking